[MINOR] Fix warnings reported by lgtm (#223)
### What changes were proposed in this pull request?

Fix warnings reported by lgtm.

### Why are the changes needed?

https://lgtm.com/projects/g/apache/incubator-uniffle/?mode=list

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

By CI.

+++ server/src/main/java/org/apache/uniffle/server/buffer/ShuffleBufferManager.java
@@ -112,7 +112,7 @@ public class ShuffleBufferManager {
     }
 
     ShuffleBuffer buffer = entry.getValue();
-    int size = buffer.append(spd);
+    long size = buffer.append(spd);
     updateSize(size, isPreAllocated);
     updateShuffleSize(appId, shuffleId, size);
     synchronized (this) {
